requirements.txt
web3==6.15.1
python-dotenv==1.0.1
rich
pandas
hyperliquid-python-sdk
